refactor: openExternalUrl を core/common へ移動し外部リンク用リスナーを共有化
- openExternalUrl は Compose に依存しない DOM 薄ラッパーであり、PageVisibility と
  同じ性質。CLAUDE.md の "core/common: Compose 非依存の純粋 KMP モジュール" の方針
  に合わせて core/ui から core/common に移動。
- JVM actual を no-op から error() に変更し fail-fast。将来 Desktop ターゲットを
  追加した際に「リンクをクリックしても無反応」というバグを silent に通さない。
- LinkInteractionListener は Compose 依存のため core/ui に
  ExternalUrlLinkInteractionListener として共有 val を新設。CreditsCard 内の
  file-private val から共有 util に切り替え、他画面でも再利用可能に。
