+++ core/common/src/wasmJsMain/kotlin/core/common/UrlOpener.wasmJs.kt
@@ -1,6 +1,6 @@
 @file:OptIn(ExperimentalWasmJsInterop::class)
 
-package core.ui.util
+package core.common
 
 @JsFun("(url) => { window.open(url, '_blank', 'noopener,noreferrer'); }")
 actual external fun openExternalUrl(url: String)
